refactor: reduce mixfiles repetition
