+++ mixfiles/nx06.exs
@@ -1,76 +1,33 @@
+Code.compile_file("project.exs")
+
 defmodule Candlex.MixProject do
   use Mix.Project
 
-  @description "An Nx backend for candle machine learning minimalist framework"
-  @source_url "https://github.com/mimiquate/candlex"
-  @version "0.1.7"
-
   def project do
-    [
-      app: :candlex,
-      description: @description,
-      version: @version,
-      elixir: "~> 1.14",
-      elixirc_paths: elixirc_paths(Mix.env()),
-      start_permanent: Mix.env() == :prod,
-      deps: deps(),
-      docs: docs(),
-      package: package(),
-      lockfile: "mixfiles/nx06.lock"
-    ]
+    Candlex.Project.project()
+    |> Keyword.put(:deps, new_deps())
+    |> Keyword.put(:lockfile, "mixfiles/nx06.lock")
   end
 
   # Run "mix help compile.app" to learn about applications.
   def application do
-    [
-      extra_applications: [:logger]
-    ]
+    Candlex.Project.application()
   end
 
-  defp elixirc_paths(:test), do: ["lib", "test/support"]
-  defp elixirc_paths(_), do: ["lib"]
-
-  # Run "mix help deps" to learn about dependencies.
-  defp deps do
-    [
-      {:nx, "~> 0.6"},
-      {:rustler_precompiled, "~> 0.7.0"},
-
-      # Optional
-      {:rustler, "~> 0.29", optional: true},
-
-      # Dev
-      {:bumblebee, "~> 0.4", only: :dev, runtime: false},
-      {:stb_image, "~> 0.6", only: :dev, runtime: false},
-      {:ex_doc, "~> 0.30.9", only: :dev, runtime: false}
-    ]
+  defp new_deps do
+    extra_deps()
+    |> Enum.reduce(
+      Candlex.Project.deps(),
+      fn extra_dep, deps ->
+        deps
+        |> List.keyreplace(elem(extra_dep,0), 0, extra_dep)
+      end
+    )
   end
 
-  defp docs do
+  defp extra_deps do
     [
-      main: "Candlex",
-      source_url: @source_url,
-      source_ref: "v#{@version}"
-    ]
-  end
-
-  defp package do
-    [
-      files: [
-        "lib",
-        "native",
-        "priv",
-        ".formatter.exs",
-        "mix.exs",
-        "CHANGELOG.md",
-        "README.md",
-        "LICENSE",
-        "checksum-*.exs"
-      ],
-      licenses: ["Apache-2.0"],
-      links: %{
-        "GitHub" => @source_url
-      }
+      {:nx, "~> 0.6.0"}
     ]
   end
 end
